Namespace shuttle vote system and random vote user
