+++ Content.Server/RoundEnd/RoundEndSystem.cs
@@ -13,8 +13,6 @@ using Content.Server.Shuttles.Components;
 using Content.Server.Shuttles.Systems;
 using Content.Server.Station.Components;
 using Content.Server.Station.Systems;
-using Content.Server.Voting;
-using Content.Server.Voting.Managers;
 using Content.Shared.Database;
 using Content.Shared.DeviceNetwork;
 using Content.Shared.GameTicking;
@@ -24,6 +22,9 @@ using Robust.Shared.Player;
 using Robust.Shared.Prototypes;
 using Robust.Shared.Timing;
 using Timer = Robust.Shared.Timing.Timer;
+
+// CD imports
+using Content.Server._CD.RoundEnd;
 
 namespace Content.Server.RoundEnd
 {
@@ -44,7 +45,8 @@ namespace Content.Server.RoundEnd
         [Dependency] private readonly EmergencyShuttleSystem _shuttle = default!;
         [Dependency] private readonly SharedAudioSystem _audio = default!;
         [Dependency] private readonly StationSystem _stationSystem = default!;
-        [Dependency] private readonly IVoteManager _voteManager = default!;
+
+        [Dependency] private readonly ShuttleVoteSystem _cdShuttleVoteSystem = default!;
 
         public TimeSpan DefaultCooldownDuration { get; set; } = TimeSpan.FromSeconds(30);
 
@@ -361,53 +363,14 @@ namespace Content.Server.RoundEnd
             {
                 if (!_shuttle.EmergencyShuttleArrived && ExpectedCountdownEnd is null)
                 {
-                    RunRestartVote();
+                    // CD: reuse this code for our shuttle vote because this would have needed to be disabled anyway
+                    _cdShuttleVoteSystem.RunRestartVote();
                     _autoCalledBefore = true;
                 }
 
                 // Always reset auto-call in case of a recall.
                 SetAutoCallTime();
             }
-        }
-        public void RunRestartVote()
-        {
-            var options = new VoteOptions
-            {
-                InitiatorText = Loc.GetString("shuttle-vote-user"),
-                Title = Loc.GetString("shuttle-vote-title"),
-                Options =
-                {
-                    (Loc.GetString("ui-vote-restart-yes"), "yes"),
-                    (Loc.GetString("ui-vote-restart-no"), "no"),
-                    (Loc.GetString("ui-vote-restart-abstain"), "abstain")
-                },
-                Duration = TimeSpan.FromSeconds(_cfg.GetCVar(CCVars.VoteTimerRestart))
-            };
-
-            var vote = _voteManager.CreateVote(options);
-
-            vote.OnFinished += (_, _) =>
-            {
-                var votesYes = vote.VotesPerOption["yes"];
-                var votesNo = vote.VotesPerOption["no"];
-                var total = votesYes + votesNo;
-
-                if (total > 0 && votesYes > votesNo)
-                {
-                    _adminLogger.Add(LogType.Vote, LogImpact.Medium, $"Round end shuttle vote succeded: {votesYes}/{votesNo}");
-                    // TODO: Add .loc files n make an unrecallable shuttle
-                    _chatManager.DispatchServerAnnouncement(Loc.GetString("Vote succeeded, round end shuttle enroute"));
-					// This is kinda cursed but whatever, stops a recall
-					_cfg.SetCVar(CCVars.EmergencyRecallTurningPoint, 0f);
-
-                    RequestRoundEnd(null, false, "round-end-system-shuttle-auto-called-announcement");
-                }
-                else
-                {
-                    _adminLogger.Add(LogType.Vote, LogImpact.Medium, $"Restart vote failed: {votesYes}/{votesNo}");
-                }
-            };
-
         }
     }
