+++ Resources/Locale/en-US/_CD/round-end/shuttle-vote.ftl
@@ -1,0 +1,11 @@
+shuttle-vote-title = "Call a shuttle to end the round?"
+
+shuttle-vote-user-1 = "Server"
+# Yes, this is intentional
+shuttle-vote-user-2 = shuttle-vote-user-2
+shuttle-vote-user-3 = Calls-The-Shuttle
+shuttle-vote-user-4 = A Friendly Bee
+shuttle-vote-user-5 = Shuttle Vote Bot
+shuttle-vote-user-6 = John Shuttlevote
+shuttle-vote-user-7 = An Unfriendly Bee
+# I ran out of ideas at this point
